# SPEC/SPECKIT/specs/002-authentication-and-access/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SargentNexus Authentication API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      summary: Authenticate a user with globally unique email credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '400':
          description: Malformed request body or invalid request shape
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
        '401':
          description: Invalid credentials
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Inactive account
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Account locked out
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/me:
    get:
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '401':
          description: Missing, expired, or invalid authentication
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      summary: Update the current authenticated user's name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated current user summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '400':
          description: Missing or invalid name fields
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
        '401':
          description: Missing authentication or unresolved authenticated user
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/change-password:
    post:
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          description: Malformed request body or invalid password policy
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
        '401':
          description: Invalid current password or missing authentication
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Authenticated caller is not allowed to change this password
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/users/{userId}/temporary-password:
    post:
      summary: Issue a one-time display temporary password for a user
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Temporary password issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporaryPasswordResponse'
        '401':
          description: Missing, expired, or invalid authentication
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Authenticated caller is not allowed to issue a temporary password for this user
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/register:
    post:
      summary: Self-register a user with an organization invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelfRegistrationRequest'
      responses:
        '201':
          description: User account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfRegistrationResponse'
        '400':
          description: Invalid invite code, malformed request body, or validation failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ProblemDetails:
      type: object
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          nullable: true
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginSuccessResponse:
      type: object
      required: [accessToken, expiresInSeconds, requiresPasswordChange, user]
      properties:
        accessToken:
          type: string
        expiresInSeconds:
          type: integer
        requiresPasswordChange:
          type: boolean
        user:
          $ref: '#/components/schemas/UserSummary'
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    UpdateProfileRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
    TemporaryPasswordResponse:
      type: object
      required: [temporaryPassword, mustChangePassword]
      properties:
        temporaryPassword:
          type: string
        mustChangePassword:
          type: boolean
    SelfRegistrationRequest:
      type: object
      required: [inviteCode, firstName, lastName, email, password]
      properties:
        inviteCode:
          type: string
          maxLength: 16
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
    SelfRegistrationResponse:
      type: object
      required: [userId, organizationId, email, role, status]
      properties:
        userId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
        status:
          type: string
    UserSummary:
      type: object
      required: [userId, role, firstName, lastName, email, status]
      properties:
        userId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
          nullable: true
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        status:
          type: string